Enable using lossless traits during encoding
2 casts were lost during the conversion from C++ to C#, add them back and activate using the lossless traits.
Add a benchmark to measure lossless and lossy decoding speed.

+++ src/LosslessTraits16.cs
@@ -13,7 +13,7 @@ internal sealed class LosslessTraits16(int maximumSampleValue, int nearLossless)
 
     internal override int ComputeReconstructedSample(int predictedValue, int errorValue)
     {
-        return predictedValue + errorValue;
+        return (ushort)(predictedValue + errorValue);
     }
 
     internal override int ModuloRange(int errorValue)
